Added web service to provide landing page details through URL: user receipt/landing/user/{$id}/auth/{$key} and to delete receipt Update the initial db with addition user Shorten the Auth key Added condition to validate Object ID

+++ src/com/tholix/web/ReceiptController.java
@@ -10,15 +10,18 @@ import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.ModelAndView;
 
 import org.joda.time.DateTime;
 
 import com.tholix.domain.ItemEntity;
 import com.tholix.domain.ReceiptEntity;
+import com.tholix.domain.UserProfileEntity;
 import com.tholix.service.ItemFeatureManager;
 import com.tholix.service.ItemManager;
 import com.tholix.service.ReceiptManager;
@@ -26,6 +29,7 @@ import com.tholix.service.StorageManager;
 import com.tholix.service.UserAuthenticationManager;
 import com.tholix.utils.DateUtil;
 import com.tholix.utils.PerformanceProfiling;
+import com.tholix.web.rest.Header;
 
 /**
  * @author hitender
@@ -34,7 +38,7 @@ import com.tholix.utils.PerformanceProfiling;
  */
 @Controller
 @RequestMapping(value = "/receipt")
-public class ReceiptController {
+public class ReceiptController extends BaseController {
 	private static final Logger log = Logger.getLogger(ReceiptController.class);
 
 	private static String nextPage = "/receipt";
@@ -67,15 +71,63 @@ public class ReceiptController {
 	public String delete(@ModelAttribute("receiptForm") ReceiptEntity receiptForm) {
         DateTime time = DateUtil.now();
         log.info("Delete receipt " + receiptForm.getId());
-
-		receiptForm = receiptManager.findOne(receiptForm.getId());
-		itemManager.deleteWhereReceipt(receiptForm);
-		receiptManager.delete(receiptForm);
-		storageManager.deleteObject(receiptForm.getReceiptBlobId());
-
-        PerformanceProfiling.log(this.getClass(), time, Thread.currentThread().getStackTrace()[1].getMethodName());
+        boolean task = delete(receiptForm.getId());
+        //TODO in case of failure to delete send message to USER
+        PerformanceProfiling.log(this.getClass(), time, Thread.currentThread().getStackTrace()[1].getMethodName(), task);
 		return "redirect:/landing.htm";
 	}
+
+    /**
+     * Delete receipt through REST request
+     *
+     * @param id
+     * @param profileId
+     * @param authKey
+     * @return
+     */
+    @RequestMapping(value = "/d/{id}/user/{profileId}/auth/{authKey}", method=RequestMethod.GET)
+    public @ResponseBody
+    Header deleteRest(@PathVariable String id, @PathVariable String profileId, @PathVariable String authKey) {
+        DateTime time = DateUtil.now();
+        log.info("Delete receipt " + id);
+
+        UserProfileEntity userProfile = authenticate(profileId, authKey);
+        if(userProfile != null) {
+            boolean task = delete(id);
+            Header header = Header.newInstance(authKey);
+            if(task) {
+                header.setStatus(Header.RESULT.SUCCESS);
+                header.setMessage("Deleted receipt successfully");
+                PerformanceProfiling.log(this.getClass(), time, Thread.currentThread().getStackTrace()[1].getMethodName(), true);
+                return header;
+            } else {
+                header.setStatus(Header.RESULT.FAILURE);
+                header.setMessage("Delete receipt un-successful");
+                PerformanceProfiling.log(this.getClass(), time, Thread.currentThread().getStackTrace()[1].getMethodName(), false);
+                return header;
+            }
+        } else {
+            Header header = getHeaderForProfileOrAuthFailure();
+            PerformanceProfiling.log(this.getClass(), time, Thread.currentThread().getStackTrace()[1].getMethodName(), false);
+            return header;
+        }
+    }
+
+    /**
+     * Delete a Receipt and its associated data
+     * @param id
+     */
+    private boolean delete(String id) {
+        ReceiptEntity receiptForm;
+        receiptForm = receiptManager.findOne(id);
+        if(receiptForm != null) {
+            itemManager.deleteWhereReceipt(receiptForm);
+            receiptManager.delete(receiptForm);
+            storageManager.deleteObject(receiptForm.getReceiptBlobId());
+            return true;
+        }
+        return false;
+    }
 
 	public void setUserAuthenticationManager(UserAuthenticationManager userAuthenticationManager) {
 		this.userAuthenticationManager = userAuthenticationManager;
